requirements.txt
numpy>=1.24
pyyaml>=6.0
matplotlib>=3.7
scikit-learn>=1.3
lightgbm>=4.0
tabulate>=0.9
